Fix: OAuth2 회원 가입 수정 , 리뷰 수정

+++ server/src/main/java/com/codestates/julsinsa/market/dto/MarketResponseDto.java
@@ -12,8 +12,8 @@ import lombok.Setter;
 public class MarketResponseDto {
 
     private Long marketId;
-    private Long lat;
-    private Long lng;
+    private double lat;
+    private double lng;
     private boolean choice;
     private String name;
     private String phone;
